backend/requirements.txt
# Core dependencies for FastAPI backend
fastapi[standard]>=0.128.0
sqlmodel>=0.0.22
pyjwt>=2.10.1
python-dotenv>=1.0.1
psycopg2-binary>=2.9.10
uvicorn[standard]>=0.34.0
bcrypt>=4.0.1
